Add systematic move proposer
TODO: Because ths one always starts at origin (so with two competing ones the second will always win), and it's way better than random, to test it we should create a systematic algorithm with a random cursor intializer.

Fix cursor initializer bug. Because we never actually used it, I didn't realize the cursor initializer was returning a cell, not an index to be used against the claimable cell index.  This had a few knock-on effects.